requirements.txt
Django>=4.0.1, <4.1
djangorestframework>=3.13.1, <3.14
psycopg2>=2.9.3, <2.10
drf-spectacular>=0.22.1, <0.23
pillow>=9.1.0, <9.2
uwsgi>=2.0.20, <2.1